Create Automation_Mass_Menu.yaml
Add this automation to your HA instance. 
It triggers the menu to reset "input_select.rx_stereo_menu_list"  back to "coverart" when the hamburger menu is tapped to close the left hand menu. 

If you have changed the names of the helper entities when setting up this dashboard, you will need to change them accordingly in this automation as well.